refactor(expo): extract client storage module
